requirements.txt
amqp==5.1.1
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
backcall==0.2.0
backports.zoneinfo==0.2.1;python_version<"3.9"
beautifulsoup4==4.12.2
billiard==3.6.4.0
celery==5.2.7
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
config==0.5.1
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
Django==4.2.1
django-environ==0.10.0
django-extensions==3.2.1
django-filter==23.2
djangorestframework==3.14.0
drf-yasg==1.21.5
exceptiongroup==1.1.1
executing==1.2.0
factory-boy==3.2.1
Faker==18.9.0
idna==3.4
importlib-metadata==6.6.0
inflection==0.5.1
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.12.2
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
kombu==5.2.4
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pluggy==1.0.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
pytest==7.3.1
pytest-cov==4.0.0
pytest_django==4.5.2
python-dateutil==2.8.2
pytz==2023.3
redis==4.5.5
requests==2.30.0
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.2
tomli==2.0.1
traitlets==5.9.0
typing-extensions==4.5.0
uritemplate==4.1.1
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
zipp==3.15.0
